refactor(): split out FileReplacer
